posts, pictures and comment section showing, no functionality yet
